Organize utils classes in their own files
Half of the utils were in their own files, the other half weren't.
Standardize the file structure.

Specify the parent `Utils` module for the helper methods so there can be
no question about which `JSON` constant we want to use. The wrapping of
the classes should already do this, but it's also in part communicating
it to the developer.
